improve service description, reorganize test data [CE-170]
