+++ src/test/lib/test_config.py
@@ -4,7 +4,7 @@ import pytest
 from orcidlink.lib import config
 from orcidlink.lib.utils import module_dir
 from orcidlink.model import ServiceDescription
-from test.data.utils import load_data_file
+from test.mocks.data import load_data_file
 
 config_file = load_data_file("config1.toml")
 config_file2 = load_data_file("config2.toml")
@@ -17,14 +17,14 @@ def my_config_file(fs):
     fs.create_file(
         "/kb/module/SERVICE_DESCRIPTION.toml", contents=service_description_toml
     )
-    fs.add_real_directory("/kb/module/src/test/data")
+    fs.add_real_directory("/kb/module/test/data")
     yield fs
 
 
 @pytest.fixture
 def my_config_file2(fs):
     fs.create_file("/kb/module/deploy/config.toml", contents=config_file2)
-    fs.add_real_directory("/kb/module/src/test/data")
+    fs.add_real_directory("/kb/module/test/data")
     yield fs
 
 
@@ -33,13 +33,13 @@ def test_get_config(my_config_file2):
     assert c.config().orcid.clientId == "REDACTED-CLIENT-ID"
     assert c.config().orcid.clientSecret == "REDACTED-CLIENT-SECRET"
     assert (
-        c.config().services.Auth2.url
-        == "https://ci.kbase.us/services/auth/api/V2/token"
+            c.config().services.Auth2.url
+            == "https://ci.kbase.us/services/auth/api/V2/token"
     )
 
 
 def test_get_service_description():
     value = config.get_service_description()
     assert type(value) == ServiceDescription
-    assert value.module_name == "ORCIDLink"
+    assert value.name == "ORCIDLink"
     assert value.language == "Python"
