+++ src/test/lib/test_config_module.py
@@ -1,6 +1,7 @@
 import pytest
 from orcidlink.lib import config
-from test.data.utils import load_data_file
+
+from test.mocks.data import load_data_file
 
 config_yaml = load_data_file("config1.toml")
 config_yaml2 = load_data_file("config2.toml")
@@ -9,14 +10,14 @@ config_yaml2 = load_data_file("config2.toml")
 @pytest.fixture
 def my_config_file(fs):
     fs.create_file("/kb/module/deploy/config.toml", contents=config_yaml)
-    fs.add_real_directory("/kb/module/src/test/data")
+    fs.add_real_directory("/kb/module/test/data")
     yield fs
 
 
 @pytest.fixture
 def my_config_file2(fs):
     fs.create_file("/kb/module/deploy/config.toml", contents=config_yaml2)
-    fs.add_real_directory("/kb/module/src/test/data")
+    fs.add_real_directory("/kb/module/test/data")
     yield fs
 
 
@@ -26,8 +27,8 @@ def test_get_config(my_config_file2):
     assert config.config().orcid.clientId == "REDACTED-CLIENT-ID"
     assert config.config().orcid.clientSecret == "REDACTED-CLIENT-SECRET"
     assert (
-        config.config().services.Auth2.url
-        == "https://ci.kbase.us/services/auth/api/V2/token"
+            config.config().services.Auth2.url
+            == "https://ci.kbase.us/services/auth/api/V2/token"
     )
