+++ src/test/routers/orcid/test_profile.py
@@ -8,7 +8,7 @@ from orcidlink.main import app
 from orcidlink.routers.orcid.profile import get_profile_to_ORCIDProfile
 from orcidlink.service_clients import orcid_api
 from orcidlink.storage import storage_model
-from test.data.utils import load_data_file, load_data_json, load_test_data
+from test.mocks.data import load_data_file, load_data_json, load_test_data
 from test.mocks.mock_contexts import (
     mock_auth_service,
     mock_orcid_api_service,
@@ -24,7 +24,7 @@ config_yaml = load_data_file("config1.toml")
 @pytest.fixture
 def fake_fs(fs):
     fs.create_file("/kb/module/deploy/config.toml", contents=config_yaml)
-    fs.add_real_directory("/kb/module/src/test/data")
+    fs.add_real_directory("/kb/module/test/data")
     yield fs
 
 
@@ -63,8 +63,8 @@ def test_router_profile_to_normalized():
         load_test_data("orcid", "profile-model")
     )
     assert (
-        get_profile_to_ORCIDProfile(orcid_id, raw_profile).dict()
-        == model_profile.dict()
+            get_profile_to_ORCIDProfile(orcid_id, raw_profile).dict()
+            == model_profile.dict()
     )
