+++ src/test/routers/test_root.py
@@ -3,7 +3,7 @@ from datetime import datetime, timezone
 import pytest
 from fastapi.testclient import TestClient
 from orcidlink.main import app
-from test.data.utils import load_data_file, load_data_json
+from test.mocks.data import load_data_file, load_data_json
 
 client = TestClient(app, raise_server_exceptions=False)
 
@@ -19,7 +19,7 @@ def fake_fs(fs):
         "/kb/module/SERVICE_DESCRIPTION.toml", contents=service_description_toml
     )
     fs.create_file("/kb/module/config/git-info.toml", contents=gitinfo_toml)
-    fs.add_real_directory("/kb/module/src/test/data")
+    fs.add_real_directory("/kb/module/test/data")
     yield fs
 
 
@@ -47,7 +47,7 @@ def test_main_info(fake_fs):
     assert response.status_code == 200
     result = response.json()
     service_description = result["service-description"]
-    assert "module-name" in service_description
-    assert service_description["module-name"] == "ORCIDLink"
+    assert "name" in service_description
+    assert service_description["name"] == "ORCIDLink"
     git_info = result["git-info"]
     assert git_info["author_name"] == "Foo Bar"
