+++ src/test/service_clients/test_auth.py
@@ -3,8 +3,8 @@ from orcidlink.lib.errors import ServiceError
 from orcidlink.service_clients import auth
 from orcidlink.service_clients.KBaseAuth import TokenInfo
 from orcidlink.service_clients.auth import ensure_authorization
-from test.data.utils import load_data_file
 from test.lib.test_responses import mock_services
+from test.mocks.data import load_data_file
 from test.mocks.mock_contexts import mock_auth_service, no_stderr
 
 config_yaml = load_data_file("config1.toml")
@@ -13,7 +13,7 @@ config_yaml = load_data_file("config1.toml")
 @pytest.fixture
 def fake_fs(fs):
     fs.create_file("/kb/module/deploy/config.toml", contents=config_yaml)
-    fs.add_real_directory("/kb/module/src/test/data")
+    fs.add_real_directory("/kb/module/test/data")
     yield fs
 
 
@@ -31,6 +31,6 @@ def test_ensure_authorization():
         assert isinstance(value, TokenInfo)
 
     with pytest.raises(
-        ServiceError, match="API call requires a KBase auth token"
+            ServiceError, match="API call requires a KBase auth token"
     ) as ex:
         ensure_authorization(None)
