+++ src/test/storage/test_storage_model.py
@@ -3,7 +3,8 @@ from orcidlink.lib.config import config
 from orcidlink.model import LinkRecord, LinkingSessionInitial, ORCIDAuth
 from orcidlink.storage.storage_model import storage_model
 from orcidlink.storage.storage_model_mongo import StorageModelMongo
-from test.data.utils import load_data_file, load_data_json
+
+from test.mocks.data import load_data_file, load_data_json
 
 config_yaml = load_data_file("config1.toml")
 
@@ -11,7 +12,7 @@ config_yaml = load_data_file("config1.toml")
 @pytest.fixture
 def fake_fs(fs):
     fs.create_file("/kb/module/deploy/config.toml", contents=config_yaml)
-    fs.add_real_directory("/kb/module/src/test/data")
+    fs.add_real_directory("/kb/module/test/data")
     yield fs
